.streamlit/ecommerce-ai-support-system-main/requirements.txt
# Core AI and ML libraries
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-ollama==0.1.3
ollama==0.3.2

# Database - Google BigQuery
google-cloud-bigquery==3.15.0
google-cloud-bigquery-storage==2.22.0
pandas-gbq==0.19.2

# Web Framework
streamlit==1.38.0

# Data Processing
pandas==2.2.2
numpy==1.26.4
pydantic==2.9.2
pydantic-settings==2.5.2

# Utilities
python-dotenv==1.0.1

# Logging and Monitoring
loguru==0.7.2
rich==13.8.1

# Optional: For containerization
# docker==7.1.0 